requirements.txt
alembic==0.9.3
amqp==2.2.2
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
billiard==3.5.0.3
bleach==2.0.0
blinker==1.4
celery==4.1.0
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
construct==2.5.3
cryptography==2.2.2
cssselect==1.0.3
Django==2.2
django-haystack==2.8.1
django-taggit==0.22.2
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
future==0.16.0
gunicorn==19.7.1
html5lib==0.999999999
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itsdangerous==0.24
jieba==0.39
Jinja2==2.9.6
kombu==4.1.0
lxml==4.2.1
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.0
parsel==1.4.0
pefile==2017.11.5
Pillow==5.1.0
psycopg2==2.7.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyMySQL==0.8.0
pyOpenSSL==17.5.0
pysolr==3.7.0
python-dateutil==2.6.1
python-dotenv==0.6.5
python-editor==1.0.3
python-ptrace==0.9.3
pytz==2018.4
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
sorl-thumbnail==12.5.0
SQLAlchemy==1.1.11
sqlparse==0.3.0
Twisted==17.9.0
urllib3==1.22
vine==1.1.4
visitor==0.1.3
w3lib==1.19.0
webencodings==0.5.1
Werkzeug==0.12.2
Whoosh==2.7.4
WTForms==2.1
zope.interface==4.5.0
